requirements.txt
Streamlit==1.25.0
streamlit_chat==0.1.1
python_dotenv==1.0.0
pypdf2==3.0.1
langchain==0.0.246
Pillow==9.5.0
arxiv==1.4.8